AB#7744 Refactor JwtHelper.ts (#434)
