+++ Website/src/rendering/src/helpers/JwtHelper.ts
@@ -1,5 +1,3 @@
-import { DecodedToken } from 'ordercloud-javascript-sdk';
-
 export function parseJwt(token: string, type: 'body' | 'header' = 'body'): unknown {
   const base64 = token.split('.')[type === 'header' ? 0 : 1].replace(/-/g, '+').replace(/_/g, '/');
   const decoded = JSON.parse(
@@ -13,35 +11,4 @@ export function parseJwt(token: string, type: 'body' | 'header' = 'body'): unkno
     )
   );
   return decoded;
-}
-
-// TODO: Move OrderCloud specific code in a OrderCloud specific file
-export function parseOrderCloudJwt(token: string): DecodedToken {
-  return parseJwt(token) as DecodedToken;
-}
-
-export function isTokenExpired(token: string): boolean {
-  try {
-    if (!token) {
-      return true;
-    }
-    const decoded = parseOrderCloudJwt(token);
-    const currentSeconds = Date.now() / 1000;
-    const currentSecondsWithBuffer = currentSeconds - 10;
-    return decoded.exp < currentSecondsWithBuffer;
-  } catch {
-    return true;
-  }
-}
-
-export function isAnonymousToken(token: string): boolean {
-  try {
-    if (!token) {
-      return true;
-    }
-    const decoded = parseOrderCloudJwt(token);
-    return Boolean(decoded.orderid);
-  } catch {
-    return true;
-  }
 }
